==> 3.-relations.ipynb <==
# Obtener relaciones

Basicamente extraje del bcn las relaciones, en base a que ley modifica a cual
no recuerdo al consulta exacta que use, pero era del estilo:

```
PREFIX dc: <http://purl.org/dc/elements/1.1/>
PREFIX bcnnorms: <http://datos.bcn.cl/ontologies/bcn-norms#>

SELECT ?norma ?numero ?titulo ?leycode ?isLatestVersion (GROUP_CONCAT(STR(?modifica); separator=", ") AS ?modificaciones)
WHERE {
  ?norma a bcnnorms:Norm ;
         bcnnorms:type <http://datos.bcn.cl/recurso/cl/norma/tipo#ley> ;
         bcnnorms:hasNumber ?numero ;
         dc:title ?titulo ;
         bcnnorms:leychileCode ?leycode ;
         bcnnorms:isLatestVersion ?isLatestVersion .
  OPTIONAL { ?norma bcnnorms:modifiesTo ?modifica . }
}
GROUP BY ?norma ?numero ?titulo ?leycode ?isLatestVersion

LIMIT 10
OFFSET 10
```

Donde fue ajustando el limit y offset segun correspondia donde obtuve 3 .csv con realciones de modificación.


import pandas as pd
dfs = []
files = ["relations/1.csv","relations/2.csv","relations/3.csv"]


for file in files:
    df = pd.read_csv(file)
    dfs.append(df)
df = pd.concat(dfs).reset_index(drop=True)
df.to_csv("data/relations.csv")
df.head()

# Buscar con cuales hacen match

relations = []

for lawcode in df.leycode:
    rel_to = {lawcode: set()}
    try:
        # Obtener el texto de modificaciones
        mods_str = df.query("leycode == @lawcode")["modificaciones"].iloc[0]

        # saltar nan
        if not isinstance(mods_str, str) or not mods_str.strip():
            continue

        # Limpiar y dividir por comas
        mods = [m.strip() for m in mods_str.split(",") if m.strip()]

        for m in mods:
            # Buscar las leyes asociadas
            laws_associated = df.query("norma == @m")["leycode"].tolist()

            if not laws_associated:
                # no hay relaciones
                # no se encontro la ley
                continue

            for assoc in laws_associated:
                rel_to[lawcode].add(int(assoc))  # Convertir a int

    except Exception as e:
        print(f"e rror con ley {lawcode}: {e}")

    relations.append(rel_to)

relations

# En el output:
# [{30684: {22740, 22960, 25563}} esto quiere decir 30684 modifica 22740, 22960 y 25563

# Guardo el diccionario de las relaciones en realtions.json


import json

merged = {}
for d in relations:
    merged.update(d)

def convert_sets(obj):
    if isinstance(obj, set):
        return list(obj)
    elif isinstance(obj, dict):
        return {k: convert_sets(v) for k, v in obj.items()}
    elif isinstance(obj, list):
        return [convert_sets(i) for i in obj]
    else:
        return obj

relations_json_ready = convert_sets(merged)

with open("data/relations.json", "w", encoding="utf-8") as f:
    json.dump(relations_json_ready, f, ensure_ascii=False, indent=4)
