==> 5.-embeddings.ipynb <==
# Embeddings

Basicamente obtener los embeddings iniciales del texto.

import networkx as nx
import pickle
from google.colab import drive

graph_path = "/content/drive/MyDrive/GML/proyecto/graph_law.pickle"

with open(graph_path, 'rb') as f:
    G = pickle.load(f)




Uso el modelo *all-MiniLM-L6-v2*, deben haber mejores opciones  
pero para empezar

!pip install -q sentence-transformers
from sentence_transformers import SentenceTransformer
model = SentenceTransformer('sentence-transformers/all-MiniLM-L6-v2')

import numpy as np
import networkx as nx
from tqdm import tqdm

# Prepara los textos
texts = []
nodes = []

for n, data in G.nodes(data=True):
    texto = data.get("texto", "")
    if texto.strip() == "":
        continue
    texts.append(texto)
    nodes.append(n)

# Generar embeddings (puedes ajustar batch_size si usas GPU)
embeddings = model.encode(texts, batch_size=16, show_progress_bar=True, convert_to_numpy=True)

# Asignar los embeddings al grafo
for node, emb in zip(nodes, embeddings):
    G.nodes[node]["embedding"] = emb


Guardamos el grafo con sus embeddings, es decir cada ley tiene un embedding inciial de su texto

pickle.dump(G, open('graph_law_embeddings.pickle', 'wb'))
